test case for typesafe config mapping
I was curious about how some constructs would work between
an implementation based on typesafe or commons-configuration.
For example, both support a notion of lists, but it looks like
archaius is always expecting an underlying structure that is
essentially `Map<String, String>`.

In this test case getList fails completely if a list is used
in the HOCON format. For another example, a duration can be
retrieved as a string, but not as a numeric value or duration.

What is the expected behavior? I thought the reason
`getRawProperty` returned `Object` was to have more flexibility
in passing these sort of types from an implementation like
typesafe config or commons-configuration, but it #265 changes
it to force a string.