fix javadoc for specifying defaults with Property
I find having to specify the default for every access a bit
cumbersome. In general it means that you have to have a constant
somewhere for the default and pass it in on every access.

Plus I would tend to favor the DRY principle in regards to
specifying defaults, more similar to the reasoning in:

https://github.com/typesafehub/config#how-to-handle-defaults

Further, forcing properties to be specified as part of the config
files is quite useful for debugging, because it means the value you
see in views like the karyon admin are the actual set of properties
used. With the defaults all over the place you essentially have to
read the code if a value is missing to find out what value would
get used.