fix #286, easier to customize AppConfig
This change splits out the two `@Provides` methods for the
`ConfigMapper` and `AppConfig` into a `DefaultArchaiusModule`.
The base `ArchaiusModule` expects something else to provide
those bindings.

The other advantage to this over expecting a sub-class to
override is that the user is free to inject whatever they need
into their `@Provides` for the `AppConfig`.

README has been updated to show the usage for both `ArchaiusModule`
and `DefaultArchaiusModule`.