Add a PrivateViewConfig class that allows usage of a "private" decoder or string interpolator.
The current implementation tries very hard to keep a single decoder and string interpolator in use throughout the entire configuration tree.

Because this existing behavior is probably depended on by someone, we introduce a new Config#getPrivateView method that returns an object that does NOT propagate changes to decoder and interpolator. This view is meant to be used locally by users that want a custom decoder for a specific use case.