Add Maps utility for creating appropriately sized maps
In several instances, Maps were being initialized with an initial capacity
too small to avoid a resize. Add a Maps utility with static helpers for
creating new HashMap and LinkedHashMap instances with an initial capacity
sized appropriately for the default load factor; this is similar to helpers
that already exist in Guava and JDK19.