Fix lambda generation for interfaces with inheritance hierarchy
Generating config proxies would break if the supplied interface inherited from
another interface with a default method; this was due to the wrong type being
supplied to the LambdaMetafactory in asFunction / asBiFunction. Switch to
using the type from the MethodHandle instead of the Method.

Additionally, add a test case that covers this scenario.