Fallback to reflection if building lambda instance fails.
In some Java versions (9 and 11 at least, others not tested) the call to LambdaMetafactory fails with a mysterious "invalid bytecode" error. Instead of trying to hack our way around the quirks of an already-deprecated java version, we simply fall back to calling the method via reflection. It's slower but it works.

Closes #682