Rollback the change to use lambdas in Config Proxies.
The use of lambdas to call default methods in Config proxy interfaces surfaced memory leaks in applications that leak configuration proxies.

Because the lambdas live in the JVMs metaspace, which is often configured to a much smaller max size than the heap, the leaking proxies have a much higher chance of causing OOM errors.

We are therefore rolling back this change temporarily, while we investigate a way to re-implement it in a way that reduces the issues.