Avoid class cast exceptions in array and enum converters
If a ParameterizedType was passed to ArrayTypeConverterFactory or EnumTypeConverterFactory,
a ClassCastException would be thrown when it attempted to unconditionally cast it to Class;
update the factories to handle this more gracefully by just returning Optional.empty().

Additionally, update ArrayTypeConverterFactory to handle arrays of primitives, and add
unit tests for Enums, arrays, and collections.